backend/requirements.txt
flask
flask-cors
requests
python-dotenv
flask-sqlalchemy
mysql-connector-python
